DT-493 Support running locally
