+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pollpush/config/JmsConfig.kt
@@ -8,8 +8,11 @@ import com.amazonaws.auth.BasicAWSCredentials
 import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration
 import com.amazonaws.services.sqs.AmazonSQS
 import com.amazonaws.services.sqs.AmazonSQSClientBuilder
+import com.amazonaws.services.sqs.model.CreateQueueRequest
+import com.amazonaws.services.sqs.model.QueueAttributeName
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
@@ -76,5 +79,19 @@ open class JmsConfig {
           .withEndpointConfiguration(EndpointConfiguration(serviceEndpoint, region))
           .withCredentials(AWSStaticCredentialsProvider(AnonymousAWSCredentials()))
           .build()
+
+  @Bean
+  @ConditionalOnExpression("'\${sqs.provider}'.equals('localstack')")
+  open fun queueUrl(@Autowired awsSqsClient: AmazonSQS,
+                    @Value("\${sqs.queue.name}") queueName: String,
+                    @Value("\${sqs.dlq.name}") dlqName: String): String {
+    val result = awsSqsClient.createQueue(CreateQueueRequest(dlqName))
+    val dlqArn = awsSqsClient.getQueueAttributes(result.queueUrl, listOf(QueueAttributeName.QueueArn.toString()))
+    awsSqsClient.createQueue(CreateQueueRequest(queueName).withAttributes(
+        mapOf(QueueAttributeName.RedrivePolicy.toString() to
+            """{"deadLetterTargetArn":"${dlqArn.attributes["QueueArn"]}","maxReceiveCount":"5"}""")
+    ))
+    return awsSqsClient.getQueueUrl(queueName).queueUrl
+  }
 
 }
